python/fast-api/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.14",
    "fastapi-users[sqlalchemy2]>=14.0.1",
    "fastapi-users-db-sqlalchemy>=7.0.0",
    "pip>=25.1.1",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.41",
    "uvicorn[standard]>=0.35.0",
    "asyncpg>=0.30.0",
]
